# eddist-admin/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eddist-admin",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/boards/": {
      "get": {
        "tags": [
          "boards"
        ],
        "operationId": "get_boards",
        "responses": {
          "200": {
            "description": "List boards successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "boards"
        ],
        "operationId": "create_board",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create board successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBoardInput"
                }
              }
            }
          }
        }
      }
    },
    "/boards/{board_key}/": {
      "get": {
        "tags": [
          "boards"
        ],
        "operationId": "get_board",
        "parameters": [
          {
            "name": "board_key",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get board successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "404": {
            "description": "Board not found"
          }
        }
      },
      "patch": {
        "tags": [
          "boards"
        ],
        "operationId": "edit_board",
        "parameters": [
          {
            "name": "board_key",
            "in": "path",
            "description": "Board Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditBoardInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Edit board successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          }
        }
      }
    },
...
    "/notices/": {
      "get": {
        "tags": [
          "notices"
        ],
        "operationId": "get_notices",
        "responses": {
          "200": {
            "description": "List notices successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notice"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notices"
        ],
        "operationId": "create_notice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoticeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Notice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/notices/{id}/": {
      "get": {
        "tags": [
          "notices"
        ],
        "operationId": "get_notice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notice ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get notice successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notice"
                }
              }
            }
          },
          "404": {
            "description": "Notice not found"
          }
        }
      },
      "delete": {
        "tags": [
          "notices"
        ],
        "operationId": "delete_notice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notice ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Notice deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not the notice author"
          },
          "404": {
            "description": "Notice not found"
          }
        }
      },
      "patch": {
        "tags": [
          "notices"
        ],
        "operationId": "update_notice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notice ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoticeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notice updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notice"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not the notice author"
          },
          "404": {
            "description": "Notice not found"
          }
        }
      }
    },
...
    "/terms/": {
      "get": {
        "tags": [
          "terms"
        ],
        "operationId": "get_terms",
        "responses": {
          "200": {
            "description": "Get terms successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terms"
                }
              }
            }
          },
          "404": {
            "description": "Terms not found"
          }
        }
      },
      "put": {
        "tags": [
          "terms"
        ],
        "operationId": "update_terms",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTermsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Terms updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terms"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Terms not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Board": {
        "type": "object",
        "required": [
          "id",
          "name",
          "board_key",
          "default_name",
          "thread_count"
        ],
        "properties": {
          "board_key": {
            "type": "string"
          },
          "default_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "thread_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateBoardInput": {
        "type": "object",
        "required": [
          "name",
          "board_key",
          "default_name",
          "local_rule"
        ],
        "properties": {
          "base_response_creation_span_sec": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "base_thread_creation_span_sec": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "board_key": {
            "type": "string"
          },
          "default_name": {
            "type": "string"
          },
          "force_metadent_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "local_rule": {
            "type": "string"
          },
          "max_author_name_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_email_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_response_body_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_response_body_lines": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_thread_name_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "threads_archive_cron": {
            "type": [
              "string",
              "null"
            ]
          },
          "threads_archive_trigger_thread_count": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
...
      "CreateNoticeInput": {
        "type": "object",
        "required": [
          "title",
          "slug",
          "content",
          "published_at"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "EditBoardInput": {
        "type": "object",
        "properties": {
          "base_response_creation_span_sec": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "base_thread_creation_span_sec": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "default_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "force_metadent_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "local_rule": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_author_name_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_email_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_response_body_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_response_body_lines": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "max_thread_name_byte_length": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_only": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "threads_archive_cron": {
            "type": [
              "string",
              "null"
            ]
          },
          "threads_archive_trigger_thread_count": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
...
      "Notice": {
        "type": "object",
        "description": "Notice model for API documentation",
        "required": [
          "id",
          "slug",
          "title",
          "content",
          "created_at",
          "updated_at",
          "published_at"
        ],
        "properties": {
          "author_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Terms": {
        "type": "object",
        "description": "Terms model for API documentation",
        "required": [
          "id",
          "content",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateNoticeInput": {
        "type": "object",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional custom slug. If not provided and title is updated, will be auto-generated from new title."
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateTermsInput": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  }
}